+++ src/protocols/uniswap-v3/decrease.ts
@@ -88,5 +88,6 @@ export async function decreaseLiquidity(
     amount1: event.args.amount1,
     txHash,
     gasUsed: receipt.gasUsed,
+    effectiveGasPrice: receipt.effectiveGasPrice,
   };
 }